Restructuring of processes
