+++ gui/Calibration.py
@@ -3,7 +3,8 @@ from PyQt5 import QtCore, QtWidgets
 class Calibration(QtWidgets.QWidget):
 
     def __init__(self, main):
-        super().__init__(parent=main, flags=QtCore.Qt.Window)
+        self.main = main
+        QtWidgets.QWidget.__init__(self, parent=None, flags=QtCore.Qt.Window)
 
         self.setupUi()
