+++ gui/DisplaySettings.py
@@ -4,7 +4,8 @@ from PyQt5 import QtCore, QtWidgets
 class DisplaySettings(QtWidgets.QWidget):
 
     def __init__(self, main):
-        super().__init__(parent=main, flags=QtCore.Qt.Window)
+        self.main = main
+        QtWidgets.QWidget.__init__(self, parent=None, flags=QtCore.Qt.Window)
 
         self._setupUi()
